requirements.txt
pika==0.10.0
ckanext-harvest=0.2
simplejson==3.3.1
pymongo==2.8.1
configparser==3.5.0b2
